+++ router/Cargo.toml
@@ -25,3 +25,6 @@ workspace-hack = { path = "../workspace-hack"}
 mutable_batch_lp = { path = "../mutable_batch_lp" }
 regex = "1"
 tokio = { version = "1.13", features = ["macros", "parking_lot"] }
+
+[features]
+kafka = ["write_buffer/kafka"]
